+++ Courier/src/utopia/courier/controller/read/LazyEmailHeadersView.scala
@@ -1,9 +1,8 @@
 package utopia.courier.controller.read
 
-import utopia.courier.model.EmailHeaders
 import utopia.courier.model.write.Recipients
+import utopia.courier.model.{EmailAddress, EmailHeaders}
 import utopia.flow.time.Now
-import utopia.flow.util.StringExtensions._
 
 import java.time.Instant
 
@@ -14,7 +13,7 @@ import java.time.Instant
   */
 class LazyEmailHeadersView(getSender: => String, getSubject: => String, getMessageId: => String, getSendTime: => Instant,
                            getRecipients: => Recipients = Recipients.empty, getInReplyTo: => String,
-                           getReferences: => Vector[String], geReplyTo: => String = "",
+                           getReferences: => Vector[String], geReplyTo: => Option[EmailAddress] = None,
                            override val receiveTime: Instant = Now)
 	extends EmailHeaders
 {
@@ -23,7 +22,7 @@ class LazyEmailHeadersView(getSender: => String, getSubject: => String, getMessa
 	override lazy val sender = getSender
 	override lazy val recipients = getRecipients
 	override lazy val subject = getSubject
-	override lazy val replyTo = geReplyTo.notEmpty.getOrElse(sender)
+	override lazy val replyTo = geReplyTo.getOrElse(sender)
 	override lazy val sendTime = getSendTime
 	override lazy val messageId: String = getMessageId
 	override lazy val inReplyTo: String = getInReplyTo
@@ -36,5 +35,5 @@ class LazyEmailHeadersView(getSender: => String, getSubject: => String, getMessa
 	  * @return A fully cached / read headers instance based on this view
 	  */
 	def toHeaders = EmailHeaders.incoming(sender, subject, messageId, sendTime, recipients, inReplyTo, references,
-		replyTo, receiveTime)
+		Some(replyTo).filterNot { _ == sender }, receiveTime)
 }